% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attributes.R
\name{unga_add_p_attribute_stem}
\alias{unga_add_p_attribute_stem}
\title{Add word stems as a positional attribute to UNGA}
\usage{
unga_add_p_attribute_stem(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, whether to output messages.}
}
\description{
\code{unga_add_p_attribute_stem} will add word stems as a positional attribute
}
