% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{unga_download_corpus}
\alias{unga_download_corpus}
\title{Download the UNGA corpus from web directory.}
\usage{
unga_download_corpus(corpus = "UNGA",
  webdir = "https://polmine.s3.eu-central-1.amazonaws.com/corpora/cwb/unga",
  archive = FALSE, version = NULL)
}
\arguments{
\item{corpus}{Name of the corpus to download (\code{character} vector), defaults to "UNGA".}

\item{webdir}{The (web) directory where the tarball resides.}

\item{archive}{A \code{logical}; whether an older version of the corpus from the archive should be installed instead.}

\item{version}{If \code{archive} is \code{TRUE}, a corpus version must be specified.}
}
\description{
\code{unga_download_corpus} will get a tarball with the indexed corpus
from a directory (web dir) and install the corpus into the UNGA package.
}
