% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_supplementary_data.R
\name{unga_download_lda_model}
\alias{unga_download_lda_model}
\alias{unga_download_global_vectors}
\title{Download and load supplementary data from web directory.}
\usage{

  unga_download_lda_model(model = "lda_UNGA_2019-01-23_180_2019-09-24.rds",
  webdir = "https://polmine.s3.eu-central-1.amazonaws.com/corpora/cwb/unga",
  verbose = TRUE)

unga_download_global_vectors(model = NULL, webdir, verbose = TRUE)
}
\arguments{
\item{model}{Name of the model.}

\item{webdir}{(web) directory where the data file resides}

\item{verbose}{A \code{logical} value, whether to output status messages.}
}
\description{
\code{unga_download_lda_model} will download a lda topic model
from a directory (web dir) and store it in the UNGA package.
\code{unga_download_global_vectors} will download a global vectors model
from a directory (web dir) and store it in the UNGA package.
}
