% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_unga.R
\name{table_alt}
\alias{table_alt}
\title{A function to explore response alternatives in a data frame}
\usage{
table_alt(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
the output from \code{\link{return}}
}
\description{
A function to explore response alternatives in a data frame
}
\examples{
set.seed(123456)
gender<-round(runif(50,1,2))
variable_1<-round(runif(50,1,5))
variable_2<-round(runif(50,1,5))
variable_3<-round(runif(50,1,5))
df<-data.frame(gender,variable_1,variable_2,variable_3)
table_alt(df)
}
\references{
Norman M. Bradburn et al. 2004. Asking questions. 2nd revised
edition. John Wiley & Sons
}
