% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_unga.R
\name{table_non_resp}
\alias{table_non_resp}
\title{A function that return proportion of missing values in a data frame}
\usage{
table_non_resp(x)
}
\arguments{
\item{x}{data frame}
}
\value{
the output from \code{\link{return}}
}
\description{
A function that return proportion of missing values in a data frame
}
\examples{
set.seed(123456)
gender<-round(runif(50,1,2))
variable_1<-sample(c(1,2,3,4,5,6,NA),50,replace = TRUE)
variable_2<-sample(c(1,2,3,4,5,6,NA),50,replace = TRUE)
variable_3<-sample(c(1,2,3,4,5,6,NA),50,replace = TRUE)
df<-data.frame(gender,variable_1,variable_2,variable_3)
table_non_resp(df)

}
\references{
Norman M. Bradburn et al. 2004. Asking questions. 2nd revised
edition. John Wiley & Sons
}
