% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_unga.R
\name{table_unga_general}
\alias{table_unga_general}
\title{Function to cross tabulate all variables in a data frame}
\usage{
table_unga_general(expvar, x)
}
\arguments{
\item{expvar}{explanatory variable}

\item{x}{data frame}
}
\value{
the output from \code{\link{return}}
}
\description{
Function to cross tabulate all variables in a data frame
}
\examples{
#' set.seed(123456)
gender<-round(runif(50,1,2))
variable_1<-round(runif(50,1,5))
variable_2<-round(runif(50,1,5))
variable_3<-round(runif(50,1,5))
df<-data.frame(gender,variable_1,variable_2,variable_3)
table_unga_general(gender,df)
}
\references{
Norman M. Bradburn et al. 2004. Asking questions. 2nd revised
edition. John Wiley & Sons
}
